Suppress warnings about end of repeated objects / arrays.
Currently JSONGen spits out nasty errors when handling a repeated object
in an array. This is caused by the fact that we do not re-start parsing
for the next object. We supress this warning because:
     1. It is missleading
     2. It breaks CLion's test runner.

A proper fix will have to follow which corrects the behaviour of the
generator, restoring these log messages to a WARNIGN level.