Refactor generator code.
There are no functional changes as part of this commit.

This change refactors the generator code to be more flexible, in
preparation for future changes to allow merging of fields across
different instances of an object (say in a repeating group).

Additional Bool tests were added to complete coverage of the touched
Generator code.

The refactor amounts to:
  - Store types not hard-coded string in the internal map
  - Simplify array handling logic, in particular removing the
    IgnoreField method and replacing it with explicit checks on
    the parser's state.