Performance enhancements to the JSON Parser
Various performance enhancements have been made to the parsing of
 JSON objects (although further work is required to handle arrays of
 embedded objects)

  - When handling embedded objects keep a pointer to the current depth.
    This prevents the need to recurse through each layer's JSON object
  - Faster key lookup (avoid use of std::map)
  - New benchmark binaries compare performance against the native
    rapid-json DOM

The Library will now outperform (~25%) the rapidJSON dom for simple objects
(see example_dom), but is still significantly slower (~40%) when there are
several layers of embedded objects (see glossary_dom). The biggest cost
here is our Reset() - but further investigation would be required to
close the gap. It is worth noting that glossary_dom is comprised
entirely of strings, meaning we do not currently gain anything from our
pre-allocated memory (each field must be strcpy'd onto the heap...)

Improving arrays of objects requires further work. The issue is that
each new object added to the array require a new JSON object to be
allocated and initialised. This is crippling for performance.